@@ -91,8 +91,8 @@
     "lodash.camelcase": "^4.3.0",
     "lodash.debounce": "^4.0.8",
     "lz-string": "^1.5.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-instantsearch-dom": "^6.40.4",
     "regenerator-runtime": "^0.14.0",
     "unfetch": "^4.2.0"
